once again

+++ install.sh
@@ -5,6 +5,10 @@ for i in .[^.]*; do
       continue
     fi
 
-    ln -s $i /workspaces/.codespaces/.persistedshare/dotfiles/$i
+    if [[ -f "/workspaces/.codespaces/.persistedshare/dotfiles/$i" ]]; then
+      echo "symlink exists"
+    else
+      ln -s $i /workspaces/.codespaces/.persistedshare/dotfiles/$i
+    fi
     source /workspaces/.codespaces/.persistedshare/dotfiles/$i
 done
